middle of revisions, fiat buy flow readability wip
