+++ src/components/Tooltip/index.tsx
@@ -38,8 +38,8 @@ interface TooltipContentProps extends Omit<PopoverProps, 'content'> {
   onOpen?: () => void
   open?: () => void
   close?: () => void
-  // time delay before showing tooltip upon hover
-  delay?: number
+  // time delay in milliseconds before showing tooltip upon hover
+  delayBeforeShow?: number
   // whether to wrap the content in a `TooltipContainer`
   wrap?: boolean
   disableHover?: boolean // disable the hover and content display
@@ -60,9 +60,18 @@ export default function Tooltip({ text, open, close, disableHover, ...rest }: To
   )
 }
 
-function TooltipContent({ content, wrap = false, open, close, disableHover, ...rest }: TooltipContentProps) {
+function TooltipContent({
+  content,
+  wrap = false,
+  open,
+  delayBeforeShow,
+  close,
+  disableHover,
+  ...rest
+}: TooltipContentProps) {
   return (
     <Popover
+      delayBeforeShow={delayBeforeShow}
       content={
         wrap ? (
           <TooltipContainer onMouseEnter={disableHover ? noOp : open} onMouseLeave={disableHover ? noOp : close}>
@@ -119,40 +128,25 @@ export function MouseoverTooltipContent({
   children,
   onOpen: openCallback = undefined,
   disableHover,
-  delay,
+  delayBeforeShow,
   ...rest
 }: Omit<TooltipContentProps, 'show'>) {
   const [show, setShow] = useState(false)
-  // Default delay is undefined (no delay), so delayShow should initialize to be true.
-  const [delayShow, setDelayShow] = useState(delay ? false : true)
   const open = () => {
     setShow(true)
     openCallback?.()
   }
   const close = () => {
     setShow(false)
-    if (delay) setDelayShow(false)
   }
-
-  useEffect(() => {
-    if (show && delay) {
-      const tooltipTimer = setTimeout(() => {
-        setDelayShow(true)
-      }, delay)
-
-      return () => {
-        clearTimeout(tooltipTimer)
-      }
-    }
-    return
-  }, [show, delay])
 
   return (
     <TooltipContent
       {...rest}
+      delayBeforeShow={delayBeforeShow}
       open={open}
       close={close}
-      show={!disableHover && show && delayShow}
+      show={!disableHover && show}
       content={disableHover ? null : content}
     >
       <div onMouseEnter={open} onMouseLeave={close}>
